refactored installer
